+++ core/src/lang/variable.rs
@@ -23,7 +23,7 @@ use super::{environment_func::EnvironmentFunc, evaluation_context::EvaluationCon
 pub enum VariableValue {
     Integer(i64),
     Float(f64),
-    Decimal(i8, u64, u64), // sign, numerator, denominator
+    Decimal(i8, u128, u128), // sign, numerator, denominator
     Bool(bool),
     Str(String),
     Dur(TimeSpan),
@@ -160,13 +160,13 @@ impl PartialOrd for VariableValue {
             }
             (VariableValue::Integer(x), VariableValue::Decimal(_, _, _)) => {
                 let x_sign = if *x < 0 { -1 } else { 1 };
-                let x_num = if *x < 0 { (-*x) as u64 } else { *x as u64 };
+                let x_num = if *x < 0 { (-*x) as u128 } else { *x as u128 };
                 let x_den = 1;
                 VariableValue::Decimal(x_sign, x_num, x_den).partial_cmp(other)
             }
             (VariableValue::Decimal(_, _, _), VariableValue::Integer(y)) => {
                 let y_sign = if *y < 0 { -1 } else { 1 };
-                let y_num = if *y < 0 { (-*y) as u64 } else { *y as u64 };
+                let y_num = if *y < 0 { (-*y) as u128 } else { *y as u128 };
                 let y_den = 1;
                 self.partial_cmp(&VariableValue::Decimal(y_sign, y_num, y_den))
             }
@@ -612,11 +612,11 @@ impl VariableValue {
         }
     }
 
-    pub fn as_decimal(&self, clock: &Clock, frame_len: f64) -> (i8, u64, u64) {
+    pub fn as_decimal(&self, clock: &Clock, frame_len: f64) -> (i8, u128, u128) {
         match self {
             VariableValue::Integer(i) => {
                 let sign = if *i < 0 { -1 } else { 1 };
-                let num = if *i < 0 { (-*i) as u64 } else { *i as u64 };
+                let num = if *i < 0 { (-*i) as u128 } else { *i as u128 };
                 (sign, num, 1)
             }
             VariableValue::Float(f) => decimal_from_float64(*f),
@@ -632,7 +632,7 @@ impl VariableValue {
                 Ok(n) => decimal_from_float64(n),
                 Err(_) => (1, 0, 1),
             },
-            VariableValue::Dur(d) => (1, d.as_micros(clock, frame_len), 1),
+            VariableValue::Dur(d) => (1, d.as_micros(clock, frame_len) as u128, 1),
             VariableValue::Func(_) => todo!(),
             VariableValue::Map(_) => (1, 0, 1),
         }
@@ -674,7 +674,7 @@ impl VariableValue {
         match self {
             VariableValue::Integer(i) => TimeSpan::Micros(i.unsigned_abs()),
             VariableValue::Float(f) => TimeSpan::Micros((f.round() as i64).unsigned_abs()),
-            VariableValue::Decimal(_, num, den) => TimeSpan::Micros(num / den),
+            VariableValue::Decimal(_, num, den) => TimeSpan::Micros((num / den) as u64),
             VariableValue::Bool(_) => TimeSpan::Micros(0), // TODO décider comment caster booléen vers durée
             VariableValue::Str(_) => TimeSpan::Micros(0),  // TODO parser la chaîne de caractères
             VariableValue::Dur(d) => *d,
